perf: reuse combined health artifacts
